+++ packages/geoinfo/server/places.js
@@ -1,11 +1,4 @@
 
-//TODO using before instead after!!!
-/*Places.before.insert(function(userId, doc) {
-	if(K.settings.geoinfo && K.settings.geoinfo.autoupdate) {
-		console.log('Geoinfo: updatePlaceGeoinfo ', doc.name);
-		doc.geoinfo = K.Geoinfo.getFieldsByLoc(doc.loc)
-	}
-});*/
 Places.after.insert(function(userId, doc) {
 	if(K.settings.geoinfo && K.settings.geoinfo.autoupdate) {
 
@@ -24,7 +17,7 @@ Places.after.insert(function(userId, doc) {
 /*
 
 //TODO uncomment whe edit plugin support moving place
-Places.before.update(function(userId, doc, fieldNames, modifier, options) {
+Places.after.update(function(userId, doc, fieldNames, modifier, options) {
 	if(_.contains(fieldNames,'loc'))
 		doc.geoinfo = K.Geoinfo.getFieldsByLoc(doc.loc)
 });